feat: Add common resource functionality
The CLI is going to be full of very similar CRUD-like objects. We want to
represent these consistently, and provide a common interface for viewing
and modifying them.

This is done through the `Resource` interface. A developer just needs to
implement it, and then they are hooked into the rest of the CLI, and
potentially one day a TUI.

Signed-off-by: Justin Chadwell <justin@unikraft.com>